add post order c code genration and translation

+++ evaluateN.py
@@ -5,6 +5,7 @@ import multiprocessing
 import time
 import itertools
 import csv
+import utils.postOrder as po
 
 def runCbmc(timeout):
 	with open(os.devnull,'w') as fnull:
@@ -67,7 +68,10 @@ def compareProgs((c,out)):
 	os.remove('cbmc'+str(os.getpid())+'.c')
 	return ret
 
-def evaluateProg(i,p,c,ll,out,pool):
+def parsePostOrder(code):
+	return True if po.parse(code) else False
+
+def evaluateProg(i,p,c,ll,out,pool,useCbmc):
 	print '\r'+p,
 	sys.stdout.flush()
 	if len(filter(lambda x:len(x)>0,out)) == 0:
@@ -76,22 +80,30 @@ def evaluateProg(i,p,c,ll,out,pool):
 		if c in out:
 			return (i,c,ll,out,0) #identical
 		else:
-			res = pool.map(compareProgs,map(lambda x:(c,x),out))
-			for f in os.listdir('.'):
-				if f.startswith('cbmc') and f.endswith('.c'):
-					os.remove(f)
-			if 0 in res:
-				return (i,c,ll,out,1) #equivalent
-			else:
-				if 2 in res:
-					return (i,c,ll,out,3) #fail
+			if useCbmc:
+				res = pool.map(compareProgs,map(lambda x:(c,x),out))
+				for f in os.listdir('.'):
+					if f.startswith('cbmc') and f.endswith('.c'):
+						os.remove(f)
+				if 0 in res:
+					return (i,c,ll,out,1) #equivalent
 				else:
-					if 1 in res:
-						return (i,c,ll,out,2) #parse
+					if 2 in res:
+						return (i,c,ll,out,3) #fail
 					else:
-						return (i,c,ll,out,4) #timeout
+						if 1 in res:
+							return (i,c,ll,out,2) #parse
+						else:
+							return (i,c,ll,out,4) #timeout
+			else:
+				res = pool.map(parsePostOrder, out)
+				if any(res):
+					return (i, c, ll, out, 3)  # fail
+				else:
+					return (i, c, ll, out, 2)  # parse
 
-def evaluate(k,fc,fll,fout,fi=None,fs=None,ff=None,fp=None,ft=None):
+
+def evaluate(k,fc,fll,fout,useCbmc,fi=None,fs=None,ff=None,fp=None,ft=None):
 	nidentical = 0
 	nsuccess = 0
 	nfail = 0
@@ -109,7 +121,7 @@ def evaluate(k,fc,fll,fout,fi=None,fs=None,ff=None,fp=None,ft=None):
 	for i in filter(lambda x: x not in groups.keys(), range(max_len)):
 		groups[i] = []
 	pool = multiprocessing.Pool(processes=k)
-	results = map(lambda i: evaluateProg(i,str(i+1).zfill(len(str(max_len)))+'/'+str(max_len),cs[i],lls[i],groups[i],pool),range(len(cs)))
+	results = map(lambda i: evaluateProg(i,str(i+1).zfill(len(str(max_len)))+'/'+str(max_len),cs[i],lls[i],groups[i],pool,useCbmc),range(len(cs)))
 	pool.close()
 	pool.join()
 	print ''
@@ -142,21 +154,21 @@ def evaluate(k,fc,fll,fout,fi=None,fs=None,ff=None,fp=None,ft=None):
 			os.remove(f)
 	return (nidentical,nsuccess,nparse,nfail,ntimeout)
 
-def main(f,k,ext):
+def main(f,k,extIn,extRef,useCbmc):
 	with open(f+'.identical.'+str(k)+'.csv','w') as fidentical:
 		with open(f+'.equivalent.'+str(k)+'.csv','w') as fsuccess:
 			with open(f+'.fail.'+str(k)+'.csv', 'w') as ffail:
 				with open(f+'.parse.'+str(k)+'.csv', 'w') as fparse:
 					with open(f+'.timeout.'+str(k)+'.csv', 'w') as ftimeout:
-						csv.writer(fidentical).writerow(['line','c',ext]+map(lambda i:'out'+str(i),range(k)))
-						csv.writer(fsuccess).writerow(['line','c',ext]+map(lambda i:'out'+str(i),range(k)))
-						csv.writer(ffail).writerow(['line','c',ext]+map(lambda i:'out'+str(i),range(k)))
-						csv.writer(fparse).writerow(['line','c',ext]+map(lambda i:'out'+str(i),range(k)))
-						csv.writer(ftimeout).writerow(['line','c',ext]+map(lambda i:'out'+str(i),range(k)))
-						with open(f+'.corpus.c','r') as fc:
-							with open(f+'.corpus.'+ext, 'r') as fll:
+						csv.writer(fidentical).writerow(['line',extRef,extIn]+map(lambda i:'out'+str(i),range(k)))
+						csv.writer(fsuccess).writerow(['line',extRef,extIn]+map(lambda i:'out'+str(i),range(k)))
+						csv.writer(ffail).writerow(['line',extRef,extIn]+map(lambda i:'out'+str(i),range(k)))
+						csv.writer(fparse).writerow(['line',extRef,extIn]+map(lambda i:'out'+str(i),range(k)))
+						csv.writer(ftimeout).writerow(['line',extRef,extIn]+map(lambda i:'out'+str(i),range(k)))
+						with open(f+'.corpus.'+extRef,'r') as fc:
+							with open(f+'.corpus.'+extIn, 'r') as fll:
 								with open(f+'.corpus.'+str(k)+'.out', 'r') as fout:
-									(nidentical,nsuccess,nparse,nfail,ntimeout) = evaluate(k,fc,fll,fout,csv.writer(fidentical),csv.writer(fsuccess),csv.writer(ffail),csv.writer(fparse),csv.writer(ftimeout))
+									(nidentical,nsuccess,nparse,nfail,ntimeout) = evaluate(k,fc,fll,fout,useCbmc,csv.writer(fidentical),csv.writer(fsuccess),csv.writer(ffail),csv.writer(fparse),csv.writer(ftimeout))
 	print str(nidentical)+' statements translated identically'
 	print str(nsuccess)+' statements translated equivalently'
 	print str(nparse)+' translated statements failed to parse'
@@ -168,14 +180,15 @@ if __name__ == "__main__":
 	parser = argparse.ArgumentParser(description="Evaluate dataset translations")
 	parser.add_argument('dataset', type=str, help="dataset to translate")
 	parser.add_argument('num_translations', type=int, help="number of translations in output for each input")
-	parser.add_argument('-ll', '--llvm', dest='l', help="generate LLVM code", action='count')
-	parser.add_argument('-pt', '--parse-tree', dest='p', help="generate parse tree code", action='count')
+	parser.add_argument('-ll', '--llvm', dest='l', help="evalaute LLVM code tanslation", action='count')
+	parser.add_argument('-pt', '--parse-tree', dest='p', help="evaluate parse tree code translation", action='count')
+	parser.add_argument('-po', '--post-order', dest='po', help="usee translations to post order code", action='count')
 	args = parser.parse_args()
 
 	if (not (args.l or args.p)) or (args.l and args.p):
 		parser.error('You need to exactly one input option (-ll or -pt, not both)')
 
-	main(args.dataset,args.num_translations,'ll' if args.l else 'pt')
+	main(args.dataset,args.num_translations,'ll' if args.l else 'pt','po' if args.po else 'c',False if args.po else True)
 	for f in os.listdir('.'):
 		if f.startswith('cbmc') and f.endswith('.c'):
 			os.remove(f)
